Add support for <FIELD command at default node

+++ source/efield.f
@@ -18,8 +18,8 @@ c
       module efield
       implicit none
       integer nprobes;
-      real*8, allocatable :: probes(:)
-      real*8, allocatable :: probe_mask(:)
+      integer, allocatable :: probes(:)
+      integer, allocatable :: probe_mask(:)
       real*8, allocatable :: fielde(:,:)
       real*8, allocatable :: dfieldx(:,:)
       real*8, allocatable :: dfieldy(:,:)
